requirements.txt
aiohttp==3.9.1
aiortc==1.6.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pyinstaller==6.3.0
av==11.0.0
websockets==12.0
